+++ collie/model/cold_start_matrix_factorization.py
@@ -211,10 +211,11 @@ class ColdStartModel(MultiStagePipeline):
 
     def _move_any_external_data_to_device(self):
         """Move the item buckets to the device before training."""
+        super()._move_any_external_data_to_device()
         self.hparams.item_buckets = self.hparams.item_buckets.to(self.device)
 
     def _copy_weights(self, old: nn.Embedding, new: nn.Embedding, buckets: torch.tensor) -> None:
-        new.weight.data.copy_(old.weight.data[buckets])
+        new.weight.data.copy_(old.weight.data[buckets.to(new.weight.device)])
 
     def set_stage(self, stage: str) -> None:
         """Set the stage for the model."""